# backend/api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: Foraged goods marketplace API
  version: 0.1.0
  # "|" keeps the newlines - the description renders as Markdown in the UI.
  description: |
    A marketplace for foraged goods — berries, mushrooms, herbs. Sellers post
    listings, buyers browse, message them and place orders. Payments are
    simulated; no money changes hands.

    ## Authentication

    Two tokens, with opposite properties.

    The **access token** is a signed JWT sent on every request:

        Authorization: Bearer <access_token>

    It is verified without touching the database, which also means it cannot be
    revoked - so it is short-lived (15 minutes by default, `ACCESS_TOKEN_TTL`).

    The **refresh token** is an opaque value in an HttpOnly `refresh_token`
    cookie, set by signup and login and valid for 7 days. `POST /auth/refresh`
    exchanges it for a new access token and rotates it. Only its hash is
    stored, so a database leak cannot resume a session.

    When an access token expires, the 401 carries
    `WWW-Authenticate: Bearer error="invalid_token"` - the signal to refresh
    rather than to show a login screen. `GET /auth/me` returns the identity
    behind a token that is still valid, for restoring a session on page load.

    Public endpoints (browsing listings, search, images) work without one.

    ## Ids

    Every id is a **uuid v7**: the leading 48 bits are a millisecond timestamp,
...
    something to parse.

    ## Errors

...

...

...
servers:
...

...
security:
  - bearerAuth: []
  - apiKeyAuth: []

# "components" is the reusable-definitions section. Nothing in here appears in
# the docs on its own - it only shows up where something $refs it.
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        A key from `POST /me/api-keys`, for machine clients. It acts as its
        owner - same permissions, same ownership checks - and is rate limited
        per key rather than per IP.

        Keys cannot manage keys: creating, listing and revoking need a session,
        so a leaked key cannot mint a replacement and survive its own
        revocation.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # documentation only; no tool verifies the format
      description: |
        Access token issued by signup or login. Expired and missing tokens are
        both answered with 401.
  schemas:
...
    Error:
      type: object
      required: [error] # "details" is omitempty in Go, so it is not required
      properties:
        error:
          type: string
          description: |
            Safe to show a user on 4xx, and written for display: sentence case,
            no trailing full stop. Render it as-is rather than transforming it.
            On 5xx this is always "Something went wrong" — the cause is in the
            server log, tagged with the `X-Request-ID` on the response.
          examples: ["Listing not found"]
        details:
          type: object
          # "any key, string value" - the YAML spelling of map[string]string.
          additionalProperties:
            type: string
          description: Reserved for field-level validation errors; absent today.
...
    CreateAPIKey:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 60
          description: What this key is for. Trimmed; must not be blank.
          examples: ["ci pipeline"]
    APIKey:
      type: object
      required: [id, name, key_prefix, last_used_at, revoked_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
          description: The first characters of the key, to tell rows apart.
          examples: ["fk_live_a3f9"]
        last_used_at:
          type: [string, "null"]
          format: date-time
          description: |
            Null until the key is used. Recorded at most once a minute, so it
            is a rough "last seen" rather than a request log.
        revoked_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
    CreatedAPIKey:
      allOf:
        - $ref: "#/components/schemas/APIKey"
        - type: object
          required: [key]
          properties:
            key:
              type: string
              description: |
                Shown once and never again. Store it now.
              examples: ["fk_live_a3f9c2..."]

    # --- auth ---------------------------------------------------------------
...
    UserInfo:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          examples: ["USER"]
    AuthResponse:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
          description: |
            Send as `Authorization: Bearer <access_token>`. The refresh token
            is NOT in the body — it comes back as an HttpOnly cookie.
        user:
          $ref: "#/components/schemas/UserInfo"
...
    Presence:
      type: object
      required: [is_online]
      description: |
        `is_online` means "seen within the last 2 minutes" (twice the
        heartbeat interval, so one missed beat is not read as offline).

        `last_seen_at` is absent when the other user has switched online
        status off, and also when they have never been seen. An offline user
        who has NOT hidden their status still sends `last_seen_at`, so an
        absent timestamp does not by itself mean "offline".
      properties:
        is_online:
          type: boolean
        last_seen_at:
          # OMITTED, not null, when unknown or hidden - the Go tag is omitempty.
          type: string
          format: date-time
    ChatUser:
      type: object
      required: [id, username, presence]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        presence:
          $ref: "#/components/schemas/Presence"
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json: # the media type the body is sent as
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: |
        No token, or one that is expired or invalid. The body is the same in
        every case; `WWW-Authenticate` says which, because a client's response
        differs: refresh, or log in.
      headers:
        WWW-Authenticate:
          description: |
            `Bearer error="invalid_token", error_description="The access token expired"`
            when the token was valid but has aged out - retry through
            `POST /auth/refresh`. A bare `Bearer` for a missing, malformed or
            forged token - the client should log in. A forged token reads as
            absent on purpose.
          schema:
            type: string
          examples:
            expired:
              value: Bearer error="invalid_token", error_description="The access token expired"
            absent:
              value: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: |
        No such resource — also returned when a resource exists but the caller
        is not a participant, so its existence is not revealed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: |
        Too many requests on this API key. Only key-authenticated requests are
        limited; browser sessions are not.
      headers:
        Retry-After:
          description: Seconds until a request will be accepted again.
          schema:
            type: integer
            examples: [2]
        X-RateLimit-Limit:
          description: |
            The bucket's capacity. Sent on every key-authenticated response,
            not only on a refusal.
          schema:
            type: integer
            examples: [60]
        X-RateLimit-Remaining:
          description: Requests left in the current bucket.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Something failed on our side; the cause is in the log.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# Each key is a URL relative to the server above, so "/listings" is really
# http://localhost:8080/api/v1/listings.
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange the refresh cookie for a new access token
      operationId: refresh
      security: []
      description: |
        Reads the `refresh_token` cookie - there is no request body, and no
        bearer token is needed, since you are refreshing precisely because the
        access token is gone.

        **The refresh token is rotated.** A new one is set on the cookie and
        the old one is revoked.

        To keep two browser tabs from logging each other out, a just-rotated
        token stays usable for 30 seconds - but it rotates only once. A replay
        inside that window returns a new access token and **no** `Set-Cookie`,
        because the request that won the rotation already put the current token
        in the shared cookie jar. After the window, presenting it again is a
        401.

        The response is the same shape as signup and login, so a client sets
        its session from any of the three with one code path - and recovering
        from a 401 costs one round trip rather than two.
      responses:
        "200":
          description: A new access token, and the identity behind it.
          headers:
            Set-Cookie:
              description: |
                The rotated HttpOnly `refresh_token`, valid for 7 days. Absent
                when this request was a replay inside the grace window and did
                not win the rotation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: |
            The cookie is missing, unknown, expired, revoked by a logout, or
            rotated more than 30 seconds ago. One answer for all of them: the
            client's move is the same, and distinguishing them would reveal
            which tokens once existed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /users/{id}/follow:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [follows]
      summary: Follow a user
      operationId: followUser
      description: |
        Idempotent - following someone you already follow is still a 204. The
        edge is one-directional: this says nothing about whether they follow
        you back.
      responses:
        "204":
          description: Followed. No body.
        "400":
          description: The id is not a UUID, or it is your own - you cannot follow yourself.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [follows]
      summary: Unfollow a user
      operationId: unfollowUser
      description: |
        Idempotent, like follow: unfollowing someone you were not following is
        still a 204, because not following them is the state you asked for.
        There is no 404 - "you were not following them" is not something a
        client can act on, and it would turn a retried mutation into a spurious
        error.
      responses:
        "204":
          description: Unfollowed. No body.
        "400":
          description: The id is not a UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{id}/followers:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [follows]
      summary: List a user's followers
      operationId: getFollowers
      description: |
        Alphabetical by username, and unpaginated - deliberately, unlike
        listings: a follow list is bounded by how many people one person
        chooses to follow, while a catalogue is not. Revisit if these ever
        run to hundreds. Requires a token: a follower list is a social graph
        rather than public profile data.
      responses:
        "200":
          description: The people following that user, each with online status.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatUser"
        "400":
          description: The id is not a UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{id}/following:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [follows]
      summary: List who a user follows
      operationId: getUserFollowing
      description: |
        The mirror of `/users/{id}/followers`, and public on the same terms:
        who someone follows is no more private here than who follows them, and
        hiding one but not the other would be an asymmetry rather than a
        privacy design. `/me/following` is the same list for the authenticated
        caller, without needing to know your own id.
      responses:
        "200":
          description: The people that user follows, each with online status.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatUser"
        "400":
          description: The id is not a UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /me/following:
    get:
      tags: [follows, me]
      summary: List the people you follow
      operationId: getFollowing
      description: |
        The "friends list". Alphabetical by username, unpaginated, and each row
        carries `presence` - so a client renders online status without a
        request per person.
      responses:
        "200":
          description: The people you follow.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /me/api-keys:
    post:
      tags: [me]
      summary: Create an API key
      operationId: createAPIKey
      # Bearer only: a key must not be able to mint another key.
      security:
        - bearerAuth: []
      description: |
        Returns the key **once**. Only its SHA-256 hash is stored, so there is
        no endpoint that can show it again - if it is lost, revoke it and
        create another.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKey"
      responses:
        "201":
          description: The key, for the only time.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedAPIKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: An API key cannot manage API keys; log in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
    get:
      tags: [me]
      summary: List your API keys
      operationId: getAPIKeys
      security:
        - bearerAuth: []
      description: |
        Never includes the keys themselves - only the prefix, which is enough
        to tell one row from another.
      responses:
        "200":
          description: Your keys, newest first. Revoked ones are included.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: An API key cannot manage API keys; log in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /me/api-keys/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [me]
      summary: Revoke an API key
      operationId: revokeAPIKey
      security:
        - bearerAuth: []
      description: |
        Takes effect on the next request - the key is looked up every time, so
        there is no cache to wait out.
      responses:
        "204":
          description: Revoked. No body.
        "400":
          description: The id is not a UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: An API key cannot manage API keys; log in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: |
            No such key of yours. Someone else's key and one you already
            revoked answer the same way, so this cannot be used to discover
            which ids exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
...